face_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
insightface==0.7.3
onnxruntime==1.17.1
numpy==1.26.4
Pillow==10.2.0
python-multipart==0.0.9
